create db folder in config dir -- move connect db fn into db folder
